# inference: make `throw` block deoptimization concrete-eval friendly