@aviatesk 2023-04-09
@nanosoldier `runbenchmarks("inference", vs=":master")`