@nanosoldier 2023-04-09
[Your benchmark job](https://github.com/JuliaLang/julia/pull/49235#issuecomment-1501028699) has completed - possible performance regressions were detected. A full report can be found [here](https://github.com/JuliaCI/NanosoldierReports/blob/master/benchmark/by_hash/6c5c054_vs_1cf5091/report.md).