@nanosoldier 2023-09-20
[Your benchmark job](https://github.com/JuliaLang/julia/pull/49235#issuecomment-1727333687) has completed - no performance regressions were detected. A full report can be found [here](https://github.com/JuliaCI/NanosoldierReports/blob/master/benchmark/by_hash/c67d17f_vs_4923e95/report.md).