@aviatesk 2023-09-20
@nanosoldier `runbenchmarks("inference", vs=":master")`