@aviatesk 2023-09-20
Given that this appears to be a clear improvement, I'll go ahead and merge it once the CI checks come back clean.