@nanosoldier 2023-09-20
[Your benchmark job](https://github.com/JuliaLang/julia/pull/49235#issuecomment-1727553823) has completed - no performance regressions were detected. A full report can be found [here](https://github.com/JuliaCI/NanosoldierReports/blob/master/benchmark/by_hash/8ef9db4_vs_151ef23/report.md).