@nanosoldier 2023-09-21
[Your benchmark job](https://github.com/JuliaLang/julia/pull/49235#issuecomment-1728952216) has completed - possible performance regressions were detected. A full report can be found [here](https://github.com/JuliaCI/NanosoldierReports/blob/master/benchmark/by_hash/47ef201_vs_5fc5556/report.md).