@aviatesk 2023-09-21
@nanosoldier `runbenchmarks("inference", vs=":master")`