@aviatesk 2023-09-26
@nanosoldier `runbenchmarks("inference", vs=":master")`