@nanosoldier 2023-09-26
[Your benchmark job](https://github.com/JuliaLang/julia/pull/49235#issuecomment-1734830813) has completed - possible performance regressions were detected. A full report can be found [here](https://github.com/JuliaCI/NanosoldierReports/blob/master/benchmark/by_hash/ec1faba_vs_13d3efb/report.md).